auto_report_pipeline1.1/requirements.txt
python-pptx>=0.6.19
pypiwin32
matplotlib>=3.4.2
pandas>=1.3
click>=8.0

